LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/calc-size-interpolation-expansion-expected.txt


FAIL value at progress 0.75 in animation of "height" from "auto" to "calc-size(any, 0px)" assert_equals: expected "calc-size(auto, 0px + size * 0.25)" but got "auto"
FAIL value at progress 0.75 in animation of "height" from "0px" to "calc-size(calc-size(min-content, size + 20px), size * 2)" assert_equals: expected "calc-size(calc-size(min-content, 20px + size), 0px + size * 2 * 0.75)" but got "0px"
FAIL value at progress 0.75 in animation of "height" from "calc-size(min-content, size * 4)" to "calc-size(calc-size(min-content, size + 20px), size * 2)" assert_equals: expected "calc-size(min-content, size * 4 * 0.25 + (20px + size) * 2 * 0.75)" but got "0px"
FAIL value at progress 0.75 in animation of "width" from "calc-size(fit-content, 20px)" to "calc-size(calc-size(fit-content, 40px), size)" assert_equals: expected "calc-size(fit-content, 35px)" but got "auto"
FAIL value at progress 0.75 in animation of "width" from "calc-size(any, 20px)" to "calc-size(calc-size(fit-content, 40px), size)" assert_equals: expected "calc-size(fit-content, 35px)" but got "auto"
FAIL value at progress 0.75 in animation of "width" from "calc-size(calc-size(any, 30px), 20px)" to "calc-size(calc-size(fit-content, 40px), size)" assert_equals: expected "calc-size(fit-content, 35px)" but got "auto"
FAIL value at progress 0.75 in animation of "width" from "calc-size(fit-content, 20px)" to "calc-size(calc-size(fit-content, 3 * size + 10px), min(size + 20px, size * 2 - 30px) + size * 2 + 80px)" assert_equals: expected "calc-size(fit-content, 5px + (80px + min(10px + 20px + size * 3, -30px + (10px + size * 3) * 2) + (10px + size * 3) * 2) * 0.75)" but got "auto"
